requirements.txt
sqlalchemy
uvicorn
bcrypt
pyjwt
pydantic
psycopg2-binary
mypy
python-multipart
fastapi
python-dotenv